Minimal ext2 implementation (#2)
* refactor(error): begin use of anyhow

* feat(dev): begin of devices implementations

* refactor(error): remove anyhow

* feat: continue implementation of devices

* feat: add `from` method for Slice and remove unused starting address

* feat: continue implementation of device

* feat: add Commit object + end test for generic implemented devices

* test(dev): add file test

* chore: update `flush` function to `commit`

* feat(dev): `read_at` and `write_at` given functions

* chore(dev): change `u32` in `Address` implementation to `usize`

* feat(ext2): begin parsing of superblock

* feat(ext2): complete superblock parsing

* feat(ext2): end implementation superblock

* feat(ext2): block group descriptors + beginning of inodes

* test(ext2): add bad parsing tests

* feat(ext2): add Osd2 struct for inodes

* feat(ext2): add read data from inodes

* refactor(dev): change Address implementation

* feat(ext2): add directory parsing

* feat(ext2): implementation of regular files

* refactor(ext2): structure parsing

* refactor(ext2): change Rc to simple borrow

* refactor(fs): delete Box to use type parameters

* refactor(ext2): celled implementation

* feat(ext2): end of directory and symlinks implementation

* refactor: split efs into crates

* revert: merge all crates into one

* feat(ext2): free block allocation

* feat(ext2): minimal version of file write

* fix(ext2): add tests for I/O operations + fix block writing

* feat(io): add connections with std

* test(ext2): add basic file write tests

* fix(ext2): fully tested and correct inode write

* refactor(fs): move generics to placeholders

* feat(ext2): Celled<Ext2> implements FileSystem

* fix(ext2): symlink stored path resolution for path with length <= 60

---------

+++ src/types.rs
@@ -1,12 +1,12 @@
 //! Definitions of needed types
 //!
-//! See [the POSIX `<sys/types.h>` header](https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html) for more informations.
+//! See [the POSIX `<sys/types.h>` header](https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html) for more information.
 
 use derive_more::{Deref, DerefMut};
 
 /// Used for device IDs.
-#[derive(Debug, Clone, Copy, Deref, DerefMut)]
-pub struct Dev(pub u16);
+#[derive(Debug, Clone, Copy, Deref, DerefMut, Default)]
+pub struct Dev(pub u32);
 
 /// Used for file serial numbers.
 #[derive(Debug, Clone, Copy, Deref, DerefMut)]
@@ -30,7 +30,7 @@ pub struct Gid(pub u16);
 
 /// Used for file sizes.
 #[derive(Debug, Clone, Copy, Deref, DerefMut)]
-pub struct Off(pub i32);
+pub struct Off(pub i64);
 
 /// Used for block sizes.
 #[derive(Debug, Clone, Copy, Deref, DerefMut)]
@@ -42,7 +42,7 @@ pub struct Blkcnt(pub i32);
 
 /// Used for time in seconds.
 #[derive(Debug, Clone, Copy, Deref, DerefMut)]
-pub struct Time(pub i32);
+pub struct Time(pub i64);
 
 /// Used for time in seconds.
 ///
@@ -53,5 +53,5 @@ pub struct Timespec {
     pub tv_sec: Time,
 
     /// Nanoseconds
-    pub tv_nsec: u32,
+    pub tv_nsec: i32,
 }
